Внесла правки по замечаниям к д/з 7.10. Отрисуй меня полностью (часть 1)
