+++ js/main.js
@@ -1,3 +1,3 @@
-import './create-simular-ads.js';
+import {createSimilarAds} from './similar-ads.js';
 
-
+createSimilarAds();//это чтобы линтер не ругался
